MODSOURMAN-956: Refactor exception
