+++ mod-source-record-manager-server/src/main/java/org/folio/services/exceptions/InvalidJobProfileForFileException.java
@@ -1,5 +1,8 @@
-package org.folio.services;
+package org.folio.services.exceptions;
 
+/**
+ * This exception is thrown when an invalid job profile is selected for an uploaded file.
+ */
 public class InvalidJobProfileForFileException extends Exception {
 
   public InvalidJobProfileForFileException(String message) {
